+++ changes.md
@@ -1,5 +1,11 @@
 # Changes
 
+## Version 4.1.0: Quality of life
+
+- Each ev42 writer module instance will now publish the number of events written to file.
+- It is no longer possible to set a stop time if the previously set stop time has passed. Also added unit tests for this feature.
+- The application will now give you an approximate size of the file it is writing to, rounded up to the nearest 10 MB, in the status messages that it produces.
+
 ## Version 4.0.0: Many, many changes
 
 - Messages with the same timestamp, in the FlatBuffer, as the previous message are not written to file.